+++ exchangelib/services/get_persona.py
@@ -1,8 +1,8 @@
 from ..util import create_element, set_xml_value, MNS
-from .common import EWSService, to_item_id
+from .common import EWSAccountService, to_item_id
 
 
-class GetPersona(EWSService):
+class GetPersona(EWSAccountService):
     """MSDN: https://docs.microsoft.com/en-us/exchange/client-developer/web-service-reference/getpersona-operation"""
     SERVICE_NAME = 'GetPersona'